Add files via upload
This Portfolio contains Projects 1, 2 , and 3. 
Project 1 includes the playbooks and analysis one Network Security, along with what our task was for this project. 
Project 2 (pptx.) includes the slideshow and analysis on Intrusion Detection Systems, along with what the task for this project was and how we got to certain conclusions.
Project 3 includes the Network Presentation, Network Analysis, and Intrusion Detection System, which will explain our task for this project was.